backend/requirements-api.txt
# Backend API Dependencies
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1
aiosqlite==0.19.0
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Scraping Core
httpx==0.26.0
beautifulsoup4==4.12.3
lxml==5.1.0
playwright==1.41.2

# Data Processing
pandas==2.2.0
openpyxl==3.1.2
xlsxwriter==3.1.9

# Document Extraction (NEW)
pdfplumber==0.10.3
python-docx==1.1.0

# OCR (NEW)
pytesseract==0.3.10
Pillow==10.2.0
opencv-python==4.9.0.80

# Streaming/WebSockets (NEW)
websockets==12.0
aiohttp==3.9.3

# AI/LLM
openai==1.12.0

# Utilities
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
